Merge DAT and MOB file features into one tool.
The DAT file contains information on the map and the MOB file on the
objects on the map. There is no reason to keep them seperate when they
are always needed together.

This also requires changes to the map structure, it no longer reflects
the DAT file, but a mix of DAT and MOB data in a form that is more
suitable for a game.